+++ benchopt/cli/process_results.py
@@ -62,8 +62,10 @@ def plot(benchmark, filename=None, kinds=('suboptimality_curve',),
 
     # Plot the results.
     from benchopt.plotting import plot_benchmark
-    plot_benchmark(result_filename, benchmark, kinds=kinds, display=display,
-                   plotly=plotly, html=html)
+    return plot_benchmark(
+        result_filename, benchmark, kinds=kinds, display=display,
+        plotly=plotly, html=html
+    )
 
 
 @process_results.command(
@@ -114,10 +116,8 @@ def merge(benchmark, filenames=None, keep='last', output=None):
         output = output.with_suffix(ext)
 
     # Merge the results.
-    from benchopt.results import save_results
-    from benchopt.results.process import merge_results
-    df = merge_results(result_filenames, keep=keep)
-    save_results(df, output)
+    from benchopt.results.process import merge as merge_helper
+    merge_helper(result_filenames, keep=keep, output=output)
 
 
 @process_results.command(